Remove _with_alias methods, replace with set_name_and_get() to allow chaining

+++ simulant/renderers/batching/render_queue.cpp
@@ -96,7 +96,7 @@ void RenderQueue::insert_renderable(Renderable&& src_renderable) {
 
         // Priorities run from -250 to +250, so we need to offset the index
         auto queue_number = priority + std::abs(RENDER_PRIORITY_MIN);
-        assert(queue_number < priority_queues_.size());
+        assert(queue_number < (int) priority_queues_.size());
 
         auto& priority_queue = priority_queues_[queue_number];
         priority_queue.insert(std::move(group), std::move(idx));
